+++ ext/opentelemetry-ext-opentracing-shim/src/opentelemetry/ext/opentracing_shim/__init__.py
@@ -17,6 +17,8 @@ import logging
 import opentracing
 from deprecated import deprecated
 
+import opentelemetry.trace as trace_api
+from opentelemetry import propagators
 from opentelemetry.ext.opentracing_shim import util
 
 logger = logging.getLogger(__name__)
@@ -184,6 +186,10 @@ class TracerShim(opentracing.Tracer):
     def __init__(self, tracer):
         super().__init__(scope_manager=ScopeManagerShim(self))
         self._otel_tracer = tracer
+        self._supported_formats = (
+            opentracing.Format.TEXT_MAP,
+            opentracing.Format.HTTP_HEADERS,
+        )
 
     def unwrap(self):
         """Returns the wrapped OpenTelemetry `Tracer` object."""
@@ -226,11 +232,15 @@ class TracerShim(opentracing.Tracer):
         # Use the specified parent or the active span if possible. Otherwise,
         # use a `None` parent, which triggers the creation of a new trace.
         parent = child_of.unwrap() if child_of else None
-        span = self._otel_tracer.create_span(operation_name, parent)
 
+        links = []
         if references:
             for ref in references:
-                span.add_link(ref.referenced_context.unwrap())
+                links.append(trace_api.Link(ref.referenced_context.unwrap()))
+
+        span = self._otel_tracer.create_span(
+            operation_name, parent, links=links
+        )
 
         if tags:
             for key, value in tags.items():
@@ -249,16 +259,32 @@ class TracerShim(opentracing.Tracer):
 
     def inject(self, span_context, format, carrier):
         # pylint: disable=redefined-builtin
-        logger.warning(
-            "Calling unimplemented method inject() on class %s",
-            self.__class__.__name__,
+        # This implementation does not perform the injecting by itself but
+        # uses the configured propagators in opentelemetry.propagators.
+        # TODO: Support Format.BINARY once it is supported in
+        # opentelemetry-python.
+        if format not in self._supported_formats:
+            raise opentracing.UnsupportedFormatException
+
+        propagator = propagators.get_global_httptextformat()
+        propagator.inject(
+            span_context.unwrap(), type(carrier).__setitem__, carrier
         )
-        # TODO: Implement.
 
     def extract(self, format, carrier):
         # pylint: disable=redefined-builtin
-        logger.warning(
-            "Calling unimplemented method extract() on class %s",
-            self.__class__.__name__,
-        )
-        # TODO: Implement.
+        # This implementation does not perform the extracing by itself but
+        # uses the configured propagators in opentelemetry.propagators.
+        # TODO: Support Format.BINARY once it is supported in
+        # opentelemetry-python.
+        if format not in self._supported_formats:
+            raise opentracing.UnsupportedFormatException
+
+        def get_as_list(dict_object, key):
+            value = dict_object.get(key)
+            return [value] if value is not None else []
+
+        propagator = propagators.get_global_httptextformat()
+        otel_context = propagator.extract(get_as_list, carrier)
+
+        return SpanContextShim(otel_context)