+++ ext/opentelemetry-ext-opentracing-shim/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Implement extract and inject support for HTTP_HEADERS and TEXT_MAP formats.
+
 ## 0.2a0
 
 Released 2019-10-29